web_crawled_data_retriever/requirements.txt
streamlit==1.25.0
streamlit-extras==0.3.6
langchain==0.1.2
sentence_transformers==2.2.2
instructorembedding==1.0.1
faiss-cpu==1.7.4
python-dotenv==1.0.0
pydantic==1.10.14
pydantic_core==2.10.1
beautifulsoup4==4.12.3
html2text==2020.1.16
nest_asyncio==1.6.0
fake_useragent==1.4.0
unstructured[pdf]==0.12.4